Laser fix but from the original file.
The file you've used to update the MXM difficulty is not working with kshootmania.
I've manually made this fix from the original convert that you had (which is originally working) using the ksm editor.
It's not that accurate because but it's almost identical. I cannot copy the original because of the editor's limitations.